fix(lib): add down broker to the broker list if it is brought back on… (#28)
* fix(lib): add down broker to the broker list if it is brought back online

This is expected. However, if the broker is brought back online, the code does not add back it to the broker pool list, and logs cannot be sent to Kafka unless a Kong restart.

This PR fixes this issue. It can now add the broker back if it is available again. There is no need to restart Kong.

Main changes.

1. Retry to get full broker list.
2. Add a new broker to form a cluster. This is quite useful as single node Kafka is rarely used in production.
3. Add tests to simulate broker2 down and up.